Sync main and develop

+++ src/common/helpers/request.ts
@@ -21,8 +21,15 @@ client.interceptors.response.use(
   },
   (error: AxiosError<ValidationBag>) => {
     if (error.response?.status === 403) {
-      window.location.reload();
-      localStorage.clear();
+      const url = new URL(error.request.responseURL);
+
+      // Temporary solution. If you see this in few months,
+      // please delete.
+
+      if (!url.pathname.includes('/api/v1/live_preview')) {
+        window.location.reload();
+        localStorage.clear();
+      }
     }
 
     if (error.response?.status === 404) {
